+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - virtual licenses to the license selection dropdown
+- demo licenses to the license selection dropdown
 
 ## 0.1.0